Add constant FS_RANGE with a value of 2, since the library initializes the accelerometer in the +/- 2g range.
Add constant TEMP_OFFSET with a value of 25.  Application note AN3393 reveals the 25deg C offset for the temperature sensor (note: this is NOT in the datasheet).

Express the accelerometer values in g and show the values as decimals

Note the sensitivity of the accelerometers on the LIS3DSH is 0.06mg/LSB in the +/- 2g FS_RANGE.  The values displayed in the serial output have too much precision.

Express the temperature values in degrees Celsius.